test(anc): merge entrydiff_test into app_test
Fold TestContainerdConfigEntryDiff_NamesDifferingEntry into app_test.go next to
the sibling TestEnvValsEqualForKey_ContainerdContent* tests and drop the stray
single-test entrydiff_test.go. No coverage change.
